backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyJWT
reportlab